Zebras: Split timeframe.py into multiple modules.
