+++ src/zebras/_plot.py
@@ -1,0 +1,2 @@
+def configure_ticks(freq, length):
+    pass
