+++ src/zebras/_timeframe.py
@@ -13,11 +13,9 @@
 
 from __future__ import annotations
 
-import copy
 import dataclasses
 from operator import itemgetter
-from typing import Optional, List, NamedTuple, Union, Collection, Dict, Any
-from typing_extensions import Literal
+from typing import Optional, List, Union, Collection, Dict, Any
 
 import numpy as np
 from toolz import first, keymap, valmap, dissoc, merge, itemmap
@@ -25,330 +23,10 @@ from toolz import first, keymap, valmap, dissoc, merge, itemmap
 from gluonts import maybe
 from gluonts.itertools import pluck_attr, columns_to_rows, select
 
+from ._base import Pad, TimeBase
 from ._period import Periods, Period, period
 from ._repr import html_table
 from ._util import AxisView, pad_axis, _replace
-
-LeftOrRight = Literal["l", "r"]
-
-
-class Pad(NamedTuple):
-    """Indicator for padded values.
-
-    >>> ts = time_series([1, 2, 3]).pad(0, left=2, right=2)
-    >>> assert len(ts == 7)
-    >>> assert list(ts) == [0, 0, 1, 2, 3, 0, 0]
-    >>> assert ts._pad.left == 2
-    >>> assert ts._pad.right == 2
-
-    """
-
-    left: int = 0
-    right: int = 0
-
-    def extend(self, left, right):
-        return Pad(
-            left=max(0, self.left + left),
-            right=max(0, self.right + right),
-        )
-
-
-@dataclasses.dataclass
-class ILoc:
-    tb: TimeBase
-
-    def __getitem__(self, idx):
-        return self.tb._slice_tdim(idx)
-
-
-@dataclasses.dataclass
-class TimeView:
-    tf: TimeFrame
-
-    def __getitem__(self, idx):
-        assert isinstance(idx, slice)
-        assert idx.step is None or idx.step == 1
-
-        start = maybe.map(idx.start, self.tf.index_of)
-        stop = maybe.map(idx.stop, self.tf.index_of)
-
-        return ILoc(self.tf)[start:stop]
-
-
-class TimeBase:
-    def _slice_tdim(self, idx):
-        raise NotImplementedError
-
-    @property
-    def iloc(self):
-        return ILoc(self)
-
-    @property
-    def loc(self):
-        return TimeView(self)
-
-    @property
-    def start(self):
-        return self.iloc[0]
-
-    @property
-    def end(self):
-        return self.iloc[-1]
-
-    def head(self, count: int) -> Periods:
-        return self.iloc[:count]
-
-    def tail(self, count: int) -> Periods:
-        if count is None:
-            return self
-
-        return self.iloc[-count:]
-
-    def resize(
-        self,
-        length: Optional[int],
-        pad_value=0,
-        pad: LeftOrRight = "l",
-        skip: LeftOrRight = "r",
-    ) -> TimeFrame:
-        """Force time frame to have length ``length``.
-
-        This pads or slices the time frame, depending on whether its size is
-        smaller or bigger than the required length.
-
-        By default we pad values on the left, and skip on the right.
-        """
-        assert pad in ("l", "r")
-        assert skip in ("l", "r")
-
-        if length is None or len(self) == length:
-            return self
-
-        if len(self) < length:
-            left = right = 0
-            if pad == "l":
-                left = length - len(self)
-            else:
-                right = length - len(self)
-
-            return self.pad(pad_value, left=left, right=right)
-
-        if skip == "l":
-            return self.iloc[len(self) - length :]
-        else:
-            return self.iloc[: length - len(self)]
-
-    def index_of(self, period: Period) -> int:
-        assert self.index is not None
-
-        return self.index.index_of(period)
-
-    def __getitem__(self, idx: Union[slice, int, str]):
-        if isinstance(idx, int):
-            subtype = idx
-        else:
-            assert isinstance(idx, slice)
-            subtype = maybe.or_(idx.start, idx.stop)
-
-        if subtype is None:
-            return self
-
-        elif isinstance(subtype, int):
-            return self.iloc[idx]
-
-        elif isinstance(subtype, Period):
-            return self.loc[idx]
-
-        raise RuntimeError(f"Unsupported type {subtype}.")
-
-    def with_index(self, index):
-        return _replace(self, index=index)
-
-
-@dataclasses.dataclass(eq=False)
-class TimeSeries(TimeBase):
-    values: np.ndarray
-    index: Optional[Periods] = None
-    name: Optional[str] = None
-    tdim: int = -1
-    metadata: Optional[dict] = None
-    _pad: Pad = Pad()
-
-    def __post_init__(self):
-        assert maybe.map_or(self.index, len, len(self)) == len(self), (
-            "Index has incorrect length. "
-            f"Expected: {len(self)}, got {len(self.index)}."
-        )
-
-    def __eq__(self, other):
-        return self.values == other
-
-    def __array__(self):
-        return self.values
-
-    def __len__(self):
-        return self.values.shape[self.tdim]
-
-    def _slice_tdim(self, idx):
-        if isinstance(idx, int):
-            return AxisView(self.values, self.tdim)[idx]
-
-        start, stop, step = idx.indices(len(self))
-        assert step == 1
-
-        return _replace(
-            self,
-            values=AxisView(self.values, self.tdim)[idx],
-            index=maybe.map(self.index, itemgetter(idx)),
-            _pad=self._pad.extend(-start, stop - len(self)),
-        )
-
-    def pad(self, value, left: int = 0, right: int = 0) -> TimeSeries:
-        assert left >= 0 and right >= 0
-
-        values = pad_axis(
-            self.values,
-            axis=self.tdim,
-            left=left,
-            right=right,
-            value=value,
-        )
-
-        index = self.index
-        if self.index is not None:
-            index = self.index.prepend(left).extend(right)
-
-        return _replace(
-            self,
-            values=values,
-            index=index,
-            _pad=self._pad.extend(left, right),
-        )
-
-    @classmethod
-    def batch(cls, xs: List[TimeSeries], into=None):
-        for series in xs:
-            assert type(series) == TimeSeries
-
-        pluck = pluck_attr(xs)
-
-        tdims = set(pluck("tdim"))
-        assert len(tdims) == 1
-        tdim = first(tdims)
-        if tdim >= 0:
-            # We insert a new axis at the front, so if tdim is counting from
-            # the left (tdim is positive) we need to shift by one to the right.
-            tdim += 1
-
-        values = np.stack(pluck("values"))
-        if into is not None:
-            values = into(values)
-
-        return BatchTimeSeries(
-            values=values,
-            tdim=tdim,
-            index=pluck("index"),
-            name=pluck("name"),
-            metadata=pluck("metadata"),
-            _pad=pluck("_pad"),
-        )
-
-    def plot(self):
-        import matplotlib.pyplot as plt
-
-        if self.index is None:
-            plt.plot(self.values)
-        else:
-            plt.plot(self.index, self.values)
-
-
-@dataclasses.dataclass
-class BatchTimeSeries(TimeBase):
-    values: np.ndarray
-    index: List[Optional[Periods]]
-    name: List[Optional[str]]
-    tdim: int
-    metadata: List[Optional[dict]]
-    _pad: List[Pad]
-
-    def _slice_tdim(self, idx):
-        start, stop, step = idx.indices(len(self))
-        assert step == 1
-
-        def calc_pad(pad):
-            pad_left = max(0, pad.left - start)
-            pad_right = max(0, pad.right + stop + 1 - len(self) - pad_left)
-
-            return Pad(pad_left, pad_right)
-
-        return _replace(
-            self,
-            values=AxisView(self.values, self.tdim)[idx],
-            index=[maybe.map(index, itemgetter(idx)) for index in self.index],
-            _pad=list(map(calc_pad, self._pad)),
-        )
-
-    @property
-    def batch_size(self):
-        return len(self.values)
-
-    def __len__(self):
-        return self.values.shape[self.tdim]
-
-    def __array__(self):
-        return self.values
-
-    @property
-    def items(self):
-        return TimeSeriesItems(self)
-
-    def pad(self, value, left: int = 0, right: int = 0) -> TimeSeries:
-        assert left >= 0 and right >= 0
-
-        values = pad_axis(
-            self.values,
-            axis=self.tdim,
-            left=left,
-            right=right,
-            value=value,
-        )
-
-        def extend_index(index):
-            return index.prepend(left).extend(right)
-
-        return _replace(
-            self,
-            values=values,
-            index=[maybe.map(index_, extend_index) for index_ in self.index],
-            _pad=[pad.extend(left, right) for pad in self._pad],
-        )
-
-
-@dataclasses.dataclass(repr=False)
-class TimeSeriesItems:
-    data: BatchTimeSeries
-
-    def __len__(self):
-        return self.data.batch_size
-
-    def __getitem__(self, idx):
-        tdim = self.data.tdim
-
-        if isinstance(idx, int):
-            cls = TimeSeries
-            if tdim > 0:
-                tdim -= 1
-        else:
-            cls = BatchTimeSeries
-
-        return cls(
-            values=self.data.values[idx],
-            index=self.data.index[idx],
-            name=self.data.name[idx],
-            metadata=self.data.metadata[idx],
-            _pad=self.data._pad[idx],
-            tdim=tdim,
-        )
 
 
 @dataclasses.dataclass
@@ -633,11 +311,6 @@ class TimeFrame(TimeBase):
 
         past, future = columns_to_rows(itemmap(split_item, self.columns))
 
-        # past_index = self.index[:past_length]
-        # future_index = self.index[past_length:]
-        # new_index = past_index.prepend(pad_left).tail(past_length).extend(future_length)
-        # future_index.extend(pad_right).head(future_length)
-
         return SplitFrame(
             _past=past,
             _future=future,
@@ -908,63 +581,6 @@ def time_frame(
     return tf
 
 
-def time_series(
-    values: Collection,
-    *,
-    index: Optional[Periods] = None,
-    start: Optional[Union[Period, str]] = None,
-    freq: Optional[str] = None,
-    tdim: int = -1,
-    name: Optional[str] = None,
-    metadata: Optional[Dict] = None,
-):
-    """Create a ``zebras.TimeSeries`` object that represents a time series.
-
-    Parameters
-    ----------
-    values
-        A sequence (e.g., list, numpy arrays) representing the values
-        of the time series.
-    index, optional
-        A ``zebras.Periods`` object representing timestamps.
-        Must have the same length as the `values`, by default None
-    start, optional
-        The start time represented by a string (e.g., "2023-01-01"),
-        or a ``zebras.Period`` object. An index will be constructed using
-        this start time and the specificed frequency, by default None
-    freq, optional
-        The frequency of the period, e.g, "H" for hourly, by default None
-    tdim, optional
-        The time dimension in `values`, by default -1
-    name: optional
-        A description for the time series. This will be the column names when
-        returned from a ``TimeFrame``.
-    metadata, optional
-        A dictionary of metadata associated with the time series, by default None
-
-    Returns
-    -------
-        A ``zebras.TimeSeries`` object.
-    """
-    values = np.array(values)
-
-    ts = TimeSeries(
-        values,
-        index=index,
-        tdim=tdim,
-        name=name,
-        metadata=metadata,
-    )
-
-    if ts.index is None and start is not None:
-        if freq is not None:
-            start = period(start, freq)
-
-        return ts.with_index(start.periods(len(ts)))
-
-    return ts
-
-
 def split_frame(
     full=None,
     *,
@@ -1051,3 +667,7 @@ def split_frame(
         return sf.with_index(start.periods(len(sf)))
 
     return sf
+
+
+# We defer importing `TimeSeries` to avoid circular imports.
+from ._time_series import TimeSeries  # noqa
